ml-service/requirements.txt
flask==2.2.3
flask-cors==3.0.10
numpy==1.24.2
pandas==2.0.0
scikit-learn==1.2.2
tensorflow==2.12.0
requests==2.28.2
joblib==1.2.0
python-dotenv==1.0.0
gunicorn==20.1.0 